requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Pre-commit hooks
pre-commit>=3.0.0

# Type stubs
pandas-stubs>=2.0.0
types-requests>=2.31.0
types-tqdm>=4.66.0
types-PyYAML>=6.0.0
